extract AUDIO constant
